+++ pycobertura/cli.py
@@ -51,12 +51,20 @@ def get_exit_code(differ, source):
     help='Write output to <file> instead of stdout.'
 )
 @click.option(
-    '-s', '--source', metavar='<source-dir>',
-    help='Provide path to source code directory for HTML output.'
+    '-s', '--source', metavar='<source-dir-or-zip>',
+    help='Provide path to source code directory for HTML output. The path can '
+         'also be a zip archive instead of a directory.'
 )
-def show(cobertura_file, format, output, source):
+@click.option(
+    '-p', '--source-prefix', metavar='<dir-prefix>',
+    help='For every file found in the coverage report, it will use this '
+         'prefix to lookup files on disk. This is especially useful when '
+         'the --source is a zip archive and the files were zipped under '
+         'a directory prefix that is not part of the source.',
+)
+def show(cobertura_file, format, output, source, source_prefix):
     """show coverage summary of a Cobertura report"""
-    cobertura = Cobertura(cobertura_file, base_path=source)
+    cobertura = Cobertura(cobertura_file, source=source)
     Reporter = reporters[format]
     reporter = Reporter(cobertura)
     report = reporter.generate()
@@ -83,8 +91,8 @@ generate each of the coverage reports is accessible. By default, the source
 will read from the Cobertura report and resolved relatively from the report's
 location. If the source is not accessible from the report's location, the
 options `--source1` and `--source2` are necessary to point to the source code
-directories. If the source is not available at all, pass `--no-source` but
-missing lines and source code will not be reported.
+directories (or zip archives). If the source is not available at all, pass
+`--no-source` but missing lines and source code will not be reported.
 """)
 @click.argument('cobertura_file1')
 @click.argument('cobertura_file2')
@@ -102,16 +110,25 @@ missing lines and source code will not be reported.
     help='Write output to <file> instead of stdout.'
 )
 @click.option(
-    '-s1', '--source1', metavar='<source-dir1>',
-    help='Provide path to source code directory of first Cobertura report. '
-         'This is necessary if the filename path defined in the report is not '
-         'accessible from the location of the report.'
+    '-s1', '--source1', metavar='<source-dir1-or-zip-archive>',
+    help='Provide path to source code directory or zip archive of first '
+         'Cobertura report. This is necessary if the filename path defined '
+         'in the report is not accessible from the location of the report.'
 )
 @click.option(
-    '-s2', '--source2', metavar='<source-dir2>',
-    help='Provide path to source code directory of second Cobertura report. '
-         'This is necessary if the filename path defined in the report is not '
-         'accessible from the location of the report.'
+    '-s2', '--source2', metavar='<source-dir2-or-zip-archive>',
+    help='Like --source1 but for the second coverage report of the diff.'
+)
+@click.option(
+    '-p1', '--source-prefix1', metavar='<dir-prefix1>',
+    help='For every file found in the coverage report, it will use this '
+         'prefix to lookup files on disk. This is especially useful when '
+         'the --source1 is a zip archive and the files were zipped under '
+         'a directory prefix that is not part of the source',
+)
+@click.option(
+    '-p2', '--source-prefix2', metavar='<dir-prefix2>',
+    help='Like --source-prefix1, but for applies for --source2.',
 )
 @click.option(
     '--source/--no-source', default=True,
@@ -123,10 +140,19 @@ missing lines and source code will not be reported.
 )
 def diff(
         cobertura_file1, cobertura_file2,
-        color, format, output, source1, source2, source):
+        color, format, output, source1, source2,
+        source_prefix1, source_prefix2, source):
     """compare coverage of two Cobertura reports"""
-    cobertura1 = Cobertura(cobertura_file1, base_path=source1)
-    cobertura2 = Cobertura(cobertura_file2, base_path=source2)
+    cobertura1 = Cobertura(
+        cobertura_file1,
+        source=source1,
+        source_prefix=source_prefix1
+    )
+    cobertura2 = Cobertura(
+        cobertura_file2,
+        source=source2,
+        source_prefix=source_prefix2
+    )
 
     Reporter = delta_reporters[format]
     reporter_args = [cobertura1, cobertura2]