+++ tests/test_cobertura.py
@@ -1,4 +1,5 @@
 import mock
+import pytest
 
 from .utils import make_cobertura
 
@@ -173,20 +174,6 @@ def test_total_hits_by_class_file():
             expected_total_misses[filename]
 
 
-def test_filepath():
-    base_path = 'foo/bar/baz'
-    cobertura = make_cobertura(base_path=base_path)
-    expected_filepaths = {
-        'Main.java': 'foo/bar/baz/Main.java',
-        'search/BinarySearch.java': 'foo/bar/baz/search/BinarySearch.java',
-        'search/ISortedArraySearch.java': 'foo/bar/baz/search/ISortedArraySearch.java',
-        'search/LinearSearch.java': 'foo/bar/baz/search/LinearSearch.java',
-    }
-    for filename in cobertura.files():
-        assert cobertura.filepath(filename) == \
-            expected_filepaths[filename]
-
-
 def test_class_file_source__sources_not_found():
     from pycobertura.cobertura import Line
     cobertura = make_cobertura('tests/cobertura.xml')
@@ -228,9 +215,13 @@ def test_line_statuses():
             expected_line_statuses[filename]
 
 
-def test_class_file_source__sources_found():
+@pytest.mark.parametrize("report, source, source_prefix", [
+    ("tests/dummy.source1/coverage.xml", None, None),
+    ("tests/dummy.source1/coverage.xml", "tests/", "dummy.source1/"),
+])
+def test_class_file_source__sources_found(report, source, source_prefix):
     from pycobertura.cobertura import Line
-    cobertura = make_cobertura('tests/dummy.source1/coverage.xml')
+    cobertura = make_cobertura(report, source=source, source_prefix=source_prefix)
     expected_sources = {
     'dummy/__init__.py': [],
         'dummy/dummy.py': [