+++ CHANGES.md
@@ -1,5 +1,17 @@
 # Release Notes
 
+## Unreleased
+* BACKWARDS INCOMPATIBLE: when a source file is not found in disk pycobertura
+  will now raise a `pycobertura.filesystem.FileSystem.FileNotFound` exception
+  instead of an `IOError`.
+* possibility to pass a zip archive containing the source code instead of a
+  directory
+* BACKWARDS INCOMPATIBLE: Rename keyword argument `Cobertura(base_path=None)` > `Cobertura(source=None)`
+* Introduce new keyword argument `Cobertura(source_prefix=None)`
+* Fix an `IOError` / `FileNotFound` error which happens when the same coverage report is provided twice
+  to `pycobertura diff` (diff in degraded mode) but the first code base (`--source1`) is missing a file
+  mentioned in the coverage report.
+
 ## 0.9.0 (2016-01-29)
 * The coverage report now displays the class's filename instead of the class's
   name, the latter being more subject to different interpretations by coverage